requirements.txt
flask
flask-login
flask-sqlalchemy
python-dotenv
bootstrap-flask
